pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "crosscloud-router"
version = "0.1.0"
description = "Entropy-based routing for multi-cloud ML inference using Shannon entropy of transformer attention distributions"
readme = "README.md"
license = {text = "Apache-2.0"}
requires-python = ">=3.11"
authors = [
    {name = "Austin Kuo", email = "aus.kuo03@gmail.com"},
]
keywords = [
    "machine-learning",
    "inference",
    "multi-cloud",
    "entropy",
    "attention",
    "routing",
    "transformers",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "torch>=2.2.0",
    "transformers>=4.40.0",
    "numpy>=1.26.0",
    "scipy>=1.13.0",
    "scikit-learn>=1.4.0",
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.29.0",
    "pydantic>=2.7.0",
    "httpx>=0.27.0",
    "structlog>=24.1.0",
]

[project.optional-dependencies]
cloud = [
    "google-cloud-bigquery>=3.20.0",
    "google-cloud-firestore>=2.19.0",
    "boto3>=1.34.0",
]
airflow = [
    "apache-airflow>=2.9.0",
    "apache-airflow-providers-google>=10.17.0",
    "apache-airflow-providers-amazon>=8.20.0",
]
dev = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "httpx>=0.27.0",
    "google-cloud-bigquery>=3.20.0",
]

[project.urls]
Repository = "https://github.com/austinkuo/crosscloud-orchestration"

[tool.setuptools.packages.find]
include = ["router*", "inference*", "telemetry*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"
asyncio_mode = "auto"
markers = [
    "slow: tests that load transformer model weights (require --runslow)",
]
